% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brazil_mun_shp}
\alias{brazil_mun_shp}
\title{Brazilian Municipality Shapefiles}
\format{
A data frame with 5573 rows and 16 variables:
\describe{
\item{\code{municipality_code}}{character Municipality code (IBGE)}
\item{\code{nm_mun}}{character Municipality name}
\item{\code{cd_rgi}}{character Region code}
\item{\code{nm_rgi}}{character Region name}
\item{\code{cd_rgint}}{character COLUMN_DESCRIPTION}
\item{\code{nm_rgint}}{character COLUMN_DESCRIPTION}
\item{\code{cd_uf}}{character State code}
\item{\code{nm_uf}}{character State name}
\item{\code{sigla_uf}}{character State acronym}
\item{\code{cd_regia}}{character COLUMN_DESCRIPTION}
\item{\code{nm_regia}}{character COLUMN_DESCRIPTION}
\item{\code{sigla_rg}}{character COLUMN_DESCRIPTION}
\item{\code{cd_concu}}{character COLUMN_DESCRIPTION}
\item{\code{nm_concu}}{character COLUMN_DESCRIPTION}
\item{\code{area_km2}}{double COLUMN_DESCRIPTION}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
brazil_mun_shp
}
\description{
Shapesfiles for municipal boundaries in Brazil (IBGE)
}
\details{
DETAILS
}
\keyword{datasets}
