% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_execution}
\alias{budget_execution}
\title{Budget Execution Rate}
\format{
A data frame with 3154 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{character Year}
\item{\code{budget_execution_rate}}{double Primary government expenditure (\%)}
}
}
\source{
\url{https://data360.worldbank.org/en/indicator/WB_WDI_GF_XPD_BUDG_ZS}
}
\usage{
budget_execution
}
\description{
Primary government expenditures as a proportion of original approved budget (\%)
}
\keyword{datasets}
