% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cash_surplus}
\alias{cash_surplus}
\title{Cash Surplus as Percentage of GDP}
\format{
A data frame with 2284 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{double Year}
\item{\code{cash_surplus_pct_gdp}}{double Cash Surplus/Deficit as Percentage of GDP}
}
}
\usage{
cash_surplus
}
\description{
Cash surplus or deficit is revenue (including grants) minus expense, minus net acquisition of nonfinancial assets. In the 1986 GFS manual nonfinancial assets were included under revenue and expenditure in gross terms. This cash surplus or deficit is closest to the earlier overall budget balance (still missing is lending minus repayments, which are now a financing item under net acquisition of financial assets).
}
\details{
DETAILS
}
\keyword{datasets}
