% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_income_group.R
\name{classify_income_group}
\alias{classify_income_group}
\title{Classify country according to income group}
\usage{
classify_income_group(data)
}
\arguments{
\item{data}{A dataset containing country codes}
}
\value{
A dataset with income group classification
}
\description{
Classify country according to income group
}
