% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{classify_merit_change}
\alias{classify_merit_change}
\title{Classify a change in merit}
\usage{
classify_merit_change(data)
}
\arguments{
\item{data}{}
}
\value{
A classified dataset
}
\description{
Classify a change in merit
}
