% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constitution}
\alias{constitution}
\title{Comparative Constitutions Project}
\format{
\subsection{\code{constitution}}{

A data frame with 16,717 rows and 13 columns:
\describe{
\item{country_code}{Country WB code}
\item{year}{Year}
\item{type_constitution}{Title of the constitutional document}
\item{cowcode}{Conflicts of War Code}
\item{country}{Country Name}
\item{systid}{Constitution Unique ID}
\item{systyear}{Year of constitution promulgation}
\item{evntid}{Year of constitutional event}
\item{evnttype}{Type of constitutional event}
\item{merit}{Constitution include provisions for the meritocratic recruitment of civil servants (e.g. exams or credential requirements).}
\item{merit_article}{Article of the constitution where meritocratic recruitment is enshrined}
\item{merit_comments}{Comments on meritocratic recruitment}
\item{region}{WB Region}
...
}
}
}
\source{
\url{https://comparativeconstitutionsproject.org/download-data/#}
}
\usage{
constitution
}
\description{
The Comparative Constitutions Project produces comprehensive data about the world’s constitutions.
}
\keyword{datasets}
