% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enterprise_surveys}
\alias{enterprise_surveys}
\title{Enterprise Surveys}
\format{
A data frame with 116 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{double Year}
\item{\code{management_practices_index}}{double Management Practices Index is a composite metric combining eight individual management practices indicators that assess various aspects of organizational behavior, including decision-making processes, goal-setting, and personnel evaluation.}
}
}
\usage{
enterprise_surveys
}
\description{
World Bank Enterprise Surveys (WBES) are nationally representative firm-level surveys with top managers and owners of businesses in over 160 economies, reaching 180 in upcoming years, that provide insight into many business environment topics such as access to finance, corruption, infrastructure, and performance, among others.
}
\details{
Considers management practices in the economy as a whole
}
\keyword{datasets}
