% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enterprise_surveys_procurement}
\alias{enterprise_surveys_procurement}
\title{Enterprise Surveys: Procurement Payments}
\format{
A data frame with 116 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code.}
\item{\code{year}}{double Year.}
\item{\code{share_paid_on_time}}{double National share of firms paid on time for their procurement contracts.}
\item{\code{share_paidontime_se}}{double Standard error of \code{share_paidontime}.}
\item{\code{gov_payment_deadline}}{double De jure government payment deadline, in days.}
\item{\code{stringency}}{double Stringency of government payment deadline, compared to a benchmark of 30 days.}
}
}
\usage{
enterprise_surveys_procurement
}
\description{
World Bank Enterprise Surveys (WBES) are nationally representative firm-level surveys with top managers and owners of businesses in over 160 economies, reaching 180 in upcoming years, that provide insight into many business environment topics such as access to finance, corruption, infrastructure, and performance, among others.
}
\details{
The de jure government payment deadline is based on a desk review conducted by the World Development Report team in 2025.
}
\keyword{datasets}
