% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosperitydata360.R
\name{fetch_and_merge_world_bank_data}
\alias{fetch_and_merge_world_bank_data}
\title{Fetch data and metadata for multiple datasets}
\usage{
fetch_and_merge_world_bank_data(dataset_ids, indicator_ids_list)
}
\arguments{
\item{dataset_ids}{A character vector of dataset ID}

\item{indicator_ids_list}{A list of indicator IDs}
}
\value{
A dataset
}
\description{
Fetch data and metadata for multiple datasets
}
