% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosperitydata360.R
\name{fetch_dataset_metadata}
\alias{fetch_dataset_metadata}
\title{Fetch dataset metadata from Prosperity Data360 API}
\usage{
fetch_dataset_metadata(dataset_ids)
}
\arguments{
\item{dataset_ids}{}
}
\value{
A metadata dataset
}
\description{
Fetch dataset metadata from Prosperity Data360 API
}
