% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosperitydata360.R
\name{fetch_indicator_metadata}
\alias{fetch_indicator_metadata}
\title{Fetch indicator metadata from Prosperity Data360 API}
\usage{
fetch_indicator_metadata(dataset_ids, indicator_ids_list)
}
\arguments{
\item{dataset_ids}{A dataset ID}

\item{indicator_ids_list}{A list of indicators}
}
\value{
A metadata dataset
}
\description{
Fetch indicator metadata from Prosperity Data360 API
}
