% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosperitydata360.R
\name{fetch_prosperitydata360_data}
\alias{fetch_prosperitydata360_data}
\title{Fetch data from Prosperity Data 360 API}
\usage{
fetch_prosperitydata360_data(dataset_id, indicator_ids)
}
\arguments{
\item{dataset_id}{Dataset ID}

\item{indicator_ids}{Indicator IDs}
}
\value{
A dataset from the Prosperity Data 360 API
}
\description{
Fetch data from Prosperity Data 360 API
}
