% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ggplot_line}
\alias{ggplot_line}
\title{Generate ggplot point and line}
\usage{
ggplot_line(data, x, y, ...)
}
\arguments{
\item{data}{A data frame}

\item{x}{X-axis variable}

\item{y}{Y-axis variable}

\item{...}{}
}
\value{
A plot
}
\description{
Generate ggplot point and line
}
