% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ggplot_scatter}
\alias{ggplot_scatter}
\title{Generate ggplot scatterplot}
\usage{
ggplot_scatter(data, x, y, ...)
}
\arguments{
\item{data}{Data}

\item{x}{X-axis variable}

\item{y}{Y-axis variable}

\item{...}{Any other arguments that can be placed in aes()}
}
\value{
A plot
}
\description{
Generate ggplot scatterplot
}
