% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{invert_merit_reversal}
\alias{invert_merit_reversal}
\title{Aggregate merit reversals}
\usage{
invert_merit_reversal(data, vars = "year")
}
\arguments{
\item{data}{A dataset}

\item{vars}{A character vector of grouping variables}
}
\value{
A dataset with the change in merit-systems
}
\description{
Aggregate merit reversals
}
