% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labor}
\alias{labor}
\title{Labor statistics}
\format{
A data frame with 6170 rows and 4 variables:
\describe{
\item{\code{country_code}}{character World Bank Country Code}
\item{\code{year}}{double Year}
\item{\code{total_labor_force}}{double Labor force, Total. Source: Education Statistics}
\item{\code{unemployment_rate}}{double Unemployment, total (\% of total labor force) (national estimate). Source: UN SDG.}
}
}
\source{
Prosperity Data 360
}
\usage{
labor
}
\description{
A combination of labor statistics from the UN SDG and the Education Statistics
}
\details{
URLs: https://prosperitydata360.worldbank.org/en/indicator/WB+EDSTATS+SL+TLF+TOTL+IN; https://prosperitydata360.worldbank.org/en/indicator/UN+SDG+SL+UEM+TOTL+NE+ZS
}
\keyword{datasets}
