% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{open_budget}
\alias{open_budget}
\title{Open Budget Survey: Legislature and Super Audit Institution Oversight Score}
\format{
A data frame with 500 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{character Year}
\item{\code{oversight_score}}{double Legislature and Super Audit Institution Oversight Score}
}
}
\usage{
open_budget
}
\description{
The role that legislatures and supreme audit institutions play in the budget process and the extent to which they are able to provide robust oversight of the budget.
}
\details{
DETAILS
}
\keyword{datasets}
