% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosperitydata360.R
\name{pivot_wider_data360}
\alias{pivot_wider_data360}
\title{Pivot Prosperity Data 360 data into longer}
\usage{
pivot_wider_data360(data)
}
\arguments{
\item{data}{}
}
\value{
A long-format dataset.
}
\description{
Pivot Prosperity Data 360 data into longer
}
