% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{procurement}
\alias{procurement}
\title{Global dataset of public procurement laws, practice, and outcomes}
\format{
A data frame with 187 rows and 15 variables:
\describe{
\item{\code{economy}}{character Economy}
\item{\code{country_code}}{character World Bank country code}
\item{\code{loggdp}}{double Logged GDP per Capita}
\item{\code{transparency_l}}{double Transparency Index (Law)}
\item{\code{transparency_p}}{double Transparency Index (Practice)}
\item{\code{competition_l}}{double Competition (Law)}
\item{\code{competition_p}}{double Competition (Practice)}
\item{\code{integrity_of_contract_l}}{double Integrity of Contract (Law)}
\item{\code{integrity_of_contract_p}}{double Integrity of Contract (Practice)}
\item{\code{limits_to_exclusion_l}}{double Limits to Exclusion (Law)}
\item{\code{limits_to_exclusion_p}}{double Limits to Exclusion (Practice)}
\item{\code{laws}}{double Laws Index}
\item{\code{practices}}{double Practices Index}
\item{\code{quality}}{double Quality Index}
\item{\code{integrity}}{double Integrity Index}
}
}
\source{
https://www.openicpsr.org/openicpsr/project/153181/version/V1/view
}
\usage{
procurement
}
\description{
All data received through surveys used in this paper was validated/reconciled through extensive readings of the relevant laws, as well as multiple follow-ups conducted by phone and/or through country visits. The data was collected by a team led by one of the authors as part of the World Bank Doing Business 2020 data collection cycle.
}
\keyword{datasets}
