% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rais_mun}
\alias{rais_mun}
\title{RAIS Municipal Dataset}
\format{
A data frame with 107076 rows and 7 variables:
\describe{
\item{\code{id_municipio}}{character Municipality identifier (IBGE code)}
\item{\code{ano}}{integer Year}
\item{\code{total_headcount}}{integer Total headcount}
\item{\code{total_new_hire}}{double Total new hires}
\item{\code{total_dismissed}}{double Total dismissals}
\item{\code{share_new_hire}}{double Share of new hires}
\item{\code{share_dismissed}}{double Share of dismissals}
}
}
\source{
: https://basedosdados.org/dataset/3e7c4d58-96ba-448e-b053-d385a829ef00?table=dabe5ea8-3bb5-4a3e-9d5a-3c7003cd4a60
}
\usage{
rais_mun
}
\description{
Summary statistics on headcount, hiring and dismissals at the municipality-year level
}
\details{
Data extracted from the Base dos Dados
}
\keyword{datasets}
