% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{scale_color_expand}
\alias{scale_color_expand}
\title{Expands color palette to number of groups}
\usage{
scale_color_expand(n_group)
}
\arguments{
\item{n_group}{Number of discrete groups}
}
\value{
An expanded scale
}
\description{
Expands color palette to number of groups
}
