% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{summarise_merit}
\alias{summarise_merit}
\title{Generate merit statistics}
\usage{
summarise_merit(data, group_var, agg_fun = "sum")
}
\arguments{
\item{data}{A data frame}

\item{group_var}{Grouping variable}

\item{agg_fun}{Type of statistics. Defaults to sum.}
}
\value{
A summary statistics of the data on merit
}
\description{
Generate merit statistics
}
