% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{summarise_merit_reversal}
\alias{summarise_merit_reversal}
\title{Aggregate merit reversals by group}
\usage{
summarise_merit_reversal(data, group_var)
}
\arguments{
\item{data}{A dataset}

\item{group_var}{A grouping variable}
}
\value{
A summarised dataset with the change in merit-systems
}
\description{
Aggregate merit reversals by group
}
