% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdi_gdp_pc}
\alias{wdi_gdp_pc}
\title{GDP per Capita}
\format{
A data frame with 6100 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{double Tear}
\item{\code{gdp_per_capita_ppp_2017}}{double GDP per capita, PPP (constant 2017 international $)}
}
}
\source{
Prosperity Data 360
}
\usage{
wdi_gdp_pc
}
\description{
GDP per capita, PPP (constant 2017 international $) is the gross domestic product converted to international dollars using purchasing power parity rates and divided by total population. It is a measure of the average economic output per person, adjusted for price level differences across countries.
}
\keyword{datasets}
