% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdr2025gov-package.R
\docType{package}
\name{wdr2025gov-package}
\alias{wdr2025gov}
\alias{wdr2025gov-package}
\title{wdr2025gov: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/galileukim/wdr2025gov}
  \item Report bugs at \url{https://github.com/galileukim/wdr2025gov/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
