% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{women_ministry}
\alias{women_ministry}
\title{Gender Statistics}
\format{
A data frame with 5035 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{double Year}
\item{\code{prop_women_ministry}}{double Proportion of women in ministerial level positions}
}
}
\source{
\url{https://data360.worldbank.org/en/indicator/WB_GS_SG_GEN_MNST_ZS}
}
\usage{
women_ministry
}
\description{
Women in ministerial level positions is the proportion of women in ministerial or equivalent positions (including deputy prime ministers) in the government. Prime Ministers/Heads of Government are included when they hold ministerial portfolios. Vice-Presidents and heads of governmental or public agencies are excluded.
}
\keyword{datasets}
