% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wwbi}
\alias{wwbi}
\title{Worldwide Bureaucracy Indicators}
\format{
A data frame with 1050 rows and 6 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{year}}{double Year}
\item{\code{share_public_sector}}{double Public sector employment, as a share of total employment}
\item{\code{income_group}}{character World Bank income group}
}
}
\source{
\url{https://prosperitydata360.worldbank.org/en/indicator/WB+WWBI+BI+EMP+TOTL+PB+ZS}
}
\usage{
wwbi
}
\description{
The Worldwide Bureaucracy Indicators (WWBI) are a unique cross-national dataset on public sector employment and wages developed by the World Bank's `Bureaucracy Lab'. They aim to help researchers, development practitioners, and policymakers gain a better understanding of the personnel dimensions of state capability, the footprint of the public sector within the overall labor market, and the fiscal implications of the public sector wage bill.
}
\keyword{datasets}
