% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wwbi_occupation}
\alias{wwbi_occupation}
\title{Worldwide Bureaucracy Indicators: Occupational Breakdown}
\format{
A data frame with 125 rows and 5 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{professional_and_technical}}{double Share of public sector employment in professional and technical occupations}
\item{\code{managerial}}{double Share of public sector employmentmanagerial occupations}
\item{\code{clerical}}{double Share of public sector employment in clerical occupations}
\item{\code{other}}{double Share of public sector employment in other occupations}
}
}
\source{
Shared by the Bureaucracy Lab.
}
\usage{
wwbi_occupation
}
\description{
A breakdown of the occupational composition of each public sector, by country. Occupational classifications are based on the International Standard Classification of Occupations (ISCO).
}
\keyword{datasets}
