% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryclass}
\alias{countryclass}
\title{World Bank Country and Lending Groups}
\format{
\subsection{\code{countryclass}}{

A data frame with 267 rows and 4 columns:
\describe{
\item{country_code}{Country WB code}
\item{economy}{Country name}
\item{region}{WB region}
\item{income_group}{WB income classification}
...
}
}
}
\source{
\url{https://comparativeconstitutionsproject.org/download-data/#}
}
\usage{
countryclass
}
\description{
This dataset is produced by the World Bank Group to classify countries as to their income levels and other groups.
}
\keyword{datasets}
