% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsps}
\alias{gsps}
\title{Global Survey of Public Servants}
\format{
A data frame with 229467 rows and 10 variables:
\describe{
\item{\code{country_code}}{World Bank country code}
\item{\code{economy}}{character Country name}
\item{\code{year}}{double Year}
\item{\code{region}}{character World Bank region}
\item{\code{income_group}}{character World Bank income group}
\item{\code{group}}{character Group of respondents}
\item{\code{mean}}{double Average for the group. See scale}
\item{\code{lower_ci}}{double Lower bound for the average}
\item{\code{upper_ci}}{double Upper bound for the average}
\item{\code{scale}}{character Scale for the average}
\item{\code{response_rate}}{double Response rate for the group}
}
}
\source{
\url{https://www.globalsurveyofpublicservants.org/data-downloads}
}
\usage{
gsps
}
\description{
This dataset is a set of surveys of public servants produced by the Bureaucracy Lab at the World Bank and partnering academic institutions.
}
\keyword{datasets}
