% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis}
\alias{reproduce_analysis}
\title{Reproduce analysis for WDR 2025, Chapter 7: Standards for Better Governance}
\usage{
reproduce_analysis()
}
\value{
The function is called for its side effects (e.g., generating plots).
}
\description{
This function sources and runs an R script to reproduce all figures contained in Chapter 7 of the WDR 2025.
All the reproduced figures are placed in the \code{figs/replication} subfolder.
}
\examples{
\dontrun{
# Run the default analysis script
reproduce_analysis()
}

}
